Refactoring and change structure
